+++ frontend/app/routes/hr-advisor/employee-profile/employment-information.tsx
@@ -6,7 +6,7 @@ import { data, Form } from 'react-router';
 import { useTranslation } from 'react-i18next';
 import * as v from 'valibot';
 
-import type { Route } from '../profile/+types/employment-information';
+import type { Route } from '../employee-profile/+types/employment-information';
 
 import type { Profile } from '~/.server/domain/models';
 import { getBranchService } from '~/.server/domain/services/branch-service';
@@ -28,8 +28,8 @@ import { InlineLink } from '~/components/links';
 import { EMPLOYEE_WFA_STATUS } from '~/domain/constants';
 import { HttpStatusCodes } from '~/errors/http-status-codes';
 import { getTranslation } from '~/i18n-config.server';
-import type { employmentInformationSchema } from '~/routes/hr-advisor/[id]/profile/validation.server';
-import { parseEmploymentInformation } from '~/routes/hr-advisor/[id]/profile/validation.server';
+import type { employmentInformationSchema } from '~/routes/hr-advisor/employee-profile/validation.server';
+import { parseEmploymentInformation } from '~/routes/hr-advisor/employee-profile/validation.server';
 import { handle as parentHandle } from '~/routes/layout';
 import { extractValidationKey } from '~/utils/validation-utils';
 
@@ -61,7 +61,7 @@ export async function action({ context, params, request }: Route.ActionArgs) {
     throw updateResult.unwrapErr();
   }
 
-  return i18nRedirect('routes/hr-advisor/[id]/profile/index.tsx', request, {
+  return i18nRedirect('routes/hr-advisor/employee-profile/index.tsx', request, {
     params: { id: currentUserId },
   });
 }
@@ -175,7 +175,7 @@ export default function EmploymentInformation({ loaderData, actionData, params }
 
   return (
     <>
-      <InlineLink className="mt-6 block" file="routes/hr-advisor/[id]/profile/index.tsx" params={params} id="back-button">
+      <InlineLink className="mt-6 block" file="routes/hr-advisor/employee-profile/index.tsx" params={params} id="back-button">
         {`< ${t('app:profile.back')}`}
       </InlineLink>
       <div className="max-w-prose">
@@ -310,7 +310,7 @@ export default function EmploymentInformation({ loaderData, actionData, params }
                   {t('app:form.save')}
                 </Button>
                 <ButtonLink
-                  file="routes/hr-advisor/[id]/profile/index.tsx"
+                  file="routes/hr-advisor/employee-profile/index.tsx"
                   params={params}
                   id="cancel-button"
                   variant="alternative"
