+++ frontend/app/routes/hr-advisor/employee-profile/index.tsx
@@ -8,7 +8,7 @@ import { faCheck, faPenToSquare, faPlus, faTriangleExclamation } from '@fortawes
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { useTranslation } from 'react-i18next';
 
-import type { Route } from '../profile/+types/index';
+import type { Route } from '../employee-profile/+types/index';
 
 import type { Profile } from '~/.server/domain/models';
 import { getCityService } from '~/.server/domain/services/city-service';
@@ -325,7 +325,7 @@ export default function EditProfile({ loaderData, params }: Route.ComponentProps
         <ProfileCard
           title={t('app:profile.personal-information.title')}
           linkLabel={t('app:profile.personal-information.link-label')}
-          file="routes/hr-advisor/[id]/profile/personal-information.tsx"
+          file="routes/hr-advisor/employee-profile/personal-information.tsx"
           isComplete={loaderData.personalInformation.isComplete}
           isNew={loaderData.personalInformation.isNew}
           params={params}
@@ -370,7 +370,7 @@ export default function EditProfile({ loaderData, params }: Route.ComponentProps
         <ProfileCard
           title={t('app:profile.employment.title')}
           linkLabel={t('app:profile.employment.link-label')}
-          file="routes/hr-advisor/[id]/profile/employment-information.tsx"
+          file="routes/hr-advisor/employee-profile/employment-information.tsx"
           isComplete={loaderData.employmentInformation.isComplete}
           isNew={loaderData.employmentInformation.isNew}
           params={params}
@@ -426,7 +426,7 @@ export default function EditProfile({ loaderData, params }: Route.ComponentProps
         <ProfileCard
           title={t('app:profile.referral.title')}
           linkLabel={t('app:profile.referral.link-label')}
-          file="routes/hr-advisor/[id]/profile/referral-preferences.tsx"
+          file="routes/hr-advisor/employee-profile/referral-preferences.tsx"
           isComplete={loaderData.referralPreferences.isComplete}
           isNew={loaderData.referralPreferences.isNew}
           params={params}
